pass errors to root provider
