+++ open-authoring/errors/withOpenAuthoringErrorHandler.tsx
@@ -1,6 +1,5 @@
 import { useEffect } from 'react'
 import { useOpenAuthoring } from '../open-authoring/OpenAuthoringProvider'
-import OpenAuthoringErrorModal from './OpenAuthoringErrorModal'
 
 declare global {
   interface Window {
@@ -17,16 +16,16 @@ export const withOpenAuthoringErrorHandler = BaseComponent => (props: {
   useEffect(() => {
     ;(async () => {
       if (props.previewError) {
-        openAuthoring.updateAuthChecks()
+        openAuthoring.setError(props.previewError)
       }
     })()
   }, [props.previewError])
 
+  if (props.previewError) {
+    return null
+  }
+
   // don't show content with initial content error
   // because the data is likely missing
-  return props.previewError ? (
-    <OpenAuthoringErrorModal error={props.previewError} />
-  ) : (
-    <BaseComponent {...props} />
-  )
+  return <BaseComponent {...props} />
 }
