+++ open-authoring/github-error/OpenAuthoringErrorModal.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState, useCallback } from 'react'
 import { useCMS } from 'tinacms'
 import OpenAuthoringError from '../OpenAuthoringError'
 import { useOpenAuthoring } from '../open-authoring/OpenAuthoringProvider'
-import { getModalProps } from '../error-actions/github-interpeter'
+import { getModalProps } from './github-interpeter'
 
 interface Props {
   error: OpenAuthoringError
@@ -18,13 +18,6 @@ const OpenAuthoringErrorModal = (props: Props) => {
   const { github } = useCMS().api
 
   const openAuthoring = useOpenAuthoring()
-
-  // Hook to update root openAuthoring state when form fails.
-  // We need to perform to check before an action is clicked (e.g start auth flow)
-  // Because if it is perform on-the-fly, the window may be blocked.
-  useEffect(() => {
-    openAuthoring.updateAuthChecks()
-  }, [])
 
   useEffect(() => {
     ;(async () => {
